Issue #134: Support for select projections against subqueries [WIP]
Motivation
----------
When using Web API OData extensions, the $select clause causes an
additional select project to be applied after the original query.  This is
not currenly supported.

Modifications
-------------
Recognize when the outer query's main from clause is a subquery, and
accept this so long as there is only a select project being applied.

Register transformers that transform any outer select projection
references that point to the subquery to instead directly reference the
expressions that built the subquery's select projection.

_groupingExpressionTransformerRegistry, which performs a similar function
for grouping, was renamed to _expressionTransformerRegistry and now works
for both purposes.

Results
-------
The query generated by $select clauses are now supported when querying
Couchbase with Web API OData extensions.

Note: While the query is now generated correct, and usuable JSON data is
returned, this is not sufficient in and of itself to support OData.  The
classes being requested in the select projection cannot be serialized into
in their default state.  A custom deserialization implementation will be
required.